add 4th dimension
hypercuuuube

this is kinda needless, but I wanted to see how tricky it might be.
the process is pretty methodical, so it wasn't too bad. I could go up to
5 dimensions without much trouble, but the file search could become very
slow not to mention it might be hard to find stuff in 5 dimensions.

one caveat is that I put the 4th dimention on the end of the location
list, so it goes depth, row, col, 4th dimension. And I make the 4th
dimension the 1000s place in the array, so it's a bit backwards in order
to preserve the map I already drew in 3 dimensions